REL: Implement Manual Refresh (#6)
* REL: Homebridge Stability

By using "PlatformAccessoryEvent" for the "identify" event name, homebridge resources were being imported into the deployed program. This can (and may have been) causing problems when starting the plugin.

* REL: Disconnected volumes remain

When a volume check completes, accessories for volumes that have become removed will remain but show as 'not reachable'

At this time, the only way to remove volumes which are not visible is to restart the plugin.

* REL: Debug Logging

Better use of debug logging via homebridge. No longer polluting the homebridge log when volume interrogation results are reported, unles debug is enabled.

* REL: Issue #3 - Mechanism to refresh

Added support for a switch that allows for the volumes to be refreshed without waiting for the polling period to expire.

* REL: Updated deployment build.

* REL: Build 0.0.8

Build prep for v0.0.8 Beta 4

Co-authored-by: Mike Price <pricemi430@yahoo.com>